[#6] 레이어 추가 & Extension 파일 생성
